+++ src/lib.rs
@@ -25,6 +25,7 @@ pub mod circuit;
 pub mod error;
 mod expr;
 pub mod gates;
+pub mod noise;
 pub mod program;
 pub mod qasm;
 pub mod rng;